requirements.txt
streamlit
openai
PyMuPDF
faiss-cpu
langchain-huggingface
sentence-transformers
numpy